@@ -37,27 +37,27 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "~8.2.0",
+    "@angular/animations": "~9.0.0-rc.9",
     "@angular/cdk": "^8.1.2",
-    "@angular/common": "~8.2.0",
-    "@angular/compiler": "~8.2.0",
-    "@angular/core": "~8.2.0",
-    "@angular/forms": "~8.2.0",
+    "@angular/common": "~9.0.0-rc.9",
+    "@angular/compiler": "~9.0.0-rc.9",
+    "@angular/core": "~9.0.0-rc.9",
+    "@angular/forms": "~9.0.0-rc.9",
     "@angular/material": "^8.1.2",
-    "@angular/platform-browser": "~8.2.0",
-    "@angular/platform-browser-dynamic": "~8.2.0",
-    "@angular/router": "~8.2.0",
+    "@angular/platform-browser": "~9.0.0-rc.9",
+    "@angular/platform-browser-dynamic": "~9.0.0-rc.9",
+    "@angular/router": "~9.0.0-rc.9",
     "normalize.css": "^8.0.1",
-    "rxjs": "~6.5.2",
+    "rxjs": "~6.5.4",
     "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.800.0",
     "@angular-devkit/build-ng-packagr": "~0.800.0",
     "@angular/cli": "~8.0.3",
-    "@angular/compiler-cli": "~8.2.0",
-    "@angular/language-service": "~8.2.0",
+    "@angular/compiler-cli": "~9.0.0-rc.9",
+    "@angular/language-service": "~9.0.0-rc.9",
     "@babel/core": "^7.5.0",
     "@storybook/addons": "^5.0.9",
     "@storybook/angular": "^5.1.9",
@@ -81,6 +81,6 @@
     "ts-node": "~7.0.0",
     "tsickle": "^0.35.0",
     "tslint": "~5.15.0",
-    "typescript": "~3.5.3"
+    "typescript": "~3.7.5"
   }
 }
