@@ -1,19 +1,20 @@
-import { Injectable, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
+import { Inject, OnInit, TemplateRef, ViewContainerRef } from '@angular/core';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { AvailableViewType } from '../_models/view-switcher.model';
 import { BoostViewSwitcherService } from '../_services/boost-view-switcher.service';
 import { Destroy$ } from '../../utils/destroy';
 
-@Injectable()
 export abstract class BoostViewSelector implements OnInit {
   @Destroy$() private readonly destroy$ = new Subject();
 
   abstract viewType: AvailableViewType;
 
-  constructor(private readonly templateRef: TemplateRef<any>,
-              private readonly viewRef: ViewContainerRef,
-              private readonly viewTypesService: BoostViewSwitcherService) {
+  constructor(
+    @Inject(TemplateRef) private readonly templateRef: TemplateRef<any>,
+    @Inject(ViewContainerRef) private readonly viewRef: ViewContainerRef,
+    @Inject(BoostViewSwitcherService) private readonly viewTypesService: BoostViewSwitcherService,
+  ) {
   }
 
   ngOnInit(): void {
